move libp2p key gen to repo
