+++ cmd/cli/id/id.go
@@ -2,15 +2,20 @@ package id
 
 import (
 	"github.com/spf13/cobra"
+	"github.com/spf13/viper"
+
+	"github.com/jedib0t/go-pretty/v6/table"
+
+	"github.com/libp2p/go-libp2p"
 
 	"github.com/bacalhau-project/bacalhau/cmd/cli/serve"
 	"github.com/bacalhau-project/bacalhau/cmd/util"
 	"github.com/bacalhau-project/bacalhau/cmd/util/flags"
 	"github.com/bacalhau-project/bacalhau/cmd/util/output"
-	"github.com/bacalhau-project/bacalhau/pkg/libp2p"
+	bac_libp2p "github.com/bacalhau-project/bacalhau/pkg/libp2p"
+	"github.com/bacalhau-project/bacalhau/pkg/repo"
 	"github.com/bacalhau-project/bacalhau/pkg/system"
 	"github.com/bacalhau-project/bacalhau/pkg/util/closer"
-	"github.com/jedib0t/go-pretty/v6/table"
 )
 
 type IDInfo struct {
@@ -56,7 +61,15 @@ var idColumns = []output.TableColumn[IDInfo]{
 }
 
 func id(cmd *cobra.Command, OS *serve.ServeOptions, outputOpts output.OutputOptions) error {
-	libp2pHost, err := libp2p.NewHost(OS.SwarmPort)
+	fsRepo, err := repo.NewFS(viper.GetString("repo"))
+	if err != nil {
+		return err
+	}
+	privKey, err := fsRepo.InitLibp2pPrivateKey(OS.SwarmPort)
+	if err != nil {
+		return err
+	}
+	libp2pHost, err := bac_libp2p.NewHost(OS.SwarmPort, libp2p.Identity(privKey))
 	if err != nil {
 		return err
 	}
