+++ pkg/node/compute.go
@@ -3,8 +3,6 @@ package node
 import (
 	"context"
 	"net/url"
-	"os"
-	"path/filepath"
 
 	"github.com/libp2p/go-libp2p/core/host"
 
@@ -17,10 +15,6 @@ import (
 	compute_publicapi "github.com/bacalhau-project/bacalhau/pkg/compute/publicapi"
 	"github.com/bacalhau-project/bacalhau/pkg/compute/sensors"
 	"github.com/bacalhau-project/bacalhau/pkg/compute/store"
-	"github.com/bacalhau-project/bacalhau/pkg/compute/store/boltdb"
-	"github.com/bacalhau-project/bacalhau/pkg/compute/store/inlocalstore"
-	"github.com/bacalhau-project/bacalhau/pkg/compute/store/inmemory"
-	"github.com/bacalhau-project/bacalhau/pkg/config"
 	"github.com/bacalhau-project/bacalhau/pkg/executor"
 	executor_util "github.com/bacalhau-project/bacalhau/pkg/executor/util"
 	"github.com/bacalhau-project/bacalhau/pkg/model"
@@ -269,36 +263,6 @@ func NewComputeNode(
 
 func (c *Compute) RegisterLocalComputeCallback(callback compute.Callback) {
 	c.computeCallback.RegisterLocalComputeCallback(callback)
-}
-
-func createExecutionStore(ctx context.Context, host host.Host) (store.ExecutionStore, error) {
-	// include the host id in the state root dir to avoid conflicts when running multiple nodes on the same machine,
-	// e.g. when running tests or when running devstack
-	configDir, err := system.EnsureConfigDir()
-	if err != nil {
-		return nil, err
-	}
-	stateRootDir := filepath.Join(configDir, "execution-state-"+host.ID().String())
-	err = os.MkdirAll(stateRootDir, os.ModePerm)
-	if err != nil {
-		return nil, err
-	}
-
-	var store store.ExecutionStore
-	storageConfig := config.GetComputeStorageConfig(host.ID().Pretty())
-	if storageConfig.StoreType == config.ExecutionStoreBoltDB {
-		store, err = boltdb.NewStore(ctx, storageConfig.Location)
-		if err != nil {
-			return nil, err
-		}
-	} else if storageConfig.StoreType == config.ExecutionStoreInMemory {
-		store = inmemory.NewStore()
-	}
-
-	return inlocalstore.NewPersistentExecutionStore(inlocalstore.PersistentJobStoreParams{
-		Store:   store,
-		RootDir: stateRootDir,
-	})
 }
 
 func (c *Compute) cleanup(ctx context.Context) {
