+++ pkg/repo/fs.go
@@ -9,14 +9,7 @@ import (
 	"github.com/mitchellh/go-homedir"
 	"github.com/rs/zerolog/log"
 
-	"github.com/bacalhau-project/bacalhau/pkg/compute/store"
-	"github.com/bacalhau-project/bacalhau/pkg/compute/store/boltdb"
-	"github.com/bacalhau-project/bacalhau/pkg/compute/store/inlocalstore"
-	memcomputestore "github.com/bacalhau-project/bacalhau/pkg/compute/store/inmemory"
 	"github.com/bacalhau-project/bacalhau/pkg/config_v2"
-	"github.com/bacalhau-project/bacalhau/pkg/jobstore"
-	boltjobstore "github.com/bacalhau-project/bacalhau/pkg/jobstore/boltdb"
-	memjobstore "github.com/bacalhau-project/bacalhau/pkg/jobstore/inmemory"
 )
 
 // cribbed from lotus
@@ -57,73 +50,6 @@ func (fsr *FsRepo) Exists() (bool, error) {
 		return false, err
 	}
 	return true, nil
-}
-
-// InitExecutionStore must be called after Init
-func (fsr *FsRepo) InitExecutionStore(ctx context.Context, prefix string) (store.ExecutionStore, error) {
-	if exists, err := fsr.Exists(); err != nil {
-		return nil, fmt.Errorf("failed to check if repo exists: %w", err)
-	} else if !exists {
-		return nil, fmt.Errorf("repo is uninitialized, cannot create ExecutionStore")
-	}
-	stateRootDir := filepath.Join(fsr.path, fmt.Sprintf("%s-execution-state", prefix))
-	if err := os.MkdirAll(stateRootDir, os.ModePerm); err != nil {
-		return nil, err
-	}
-	// load the compute nodes execution store config
-	var storeCfg config_v2.StorageConfig
-	if err := config_v2.GetConfigForKey(config_v2.NodeComputeExecutionStore, &storeCfg); err != nil {
-		return nil, err
-	}
-	var (
-		store store.ExecutionStore
-		err   error
-	)
-	switch storeCfg.Type {
-	case config_v2.BoltDB:
-		path := storeCfg.Path
-		if path == "" {
-			path = filepath.Join(stateRootDir, fmt.Sprintf("%s-compute.db", prefix))
-		}
-		store, err = boltdb.NewStore(ctx, path)
-		if err != nil {
-			return nil, err
-		}
-	case config_v2.InMemory:
-		store = memcomputestore.NewStore()
-	default:
-		return nil, fmt.Errorf("unknown JobStore type: %s", storeCfg.Type)
-	}
-	return inlocalstore.NewPersistentExecutionStore(inlocalstore.PersistentJobStoreParams{
-		Store:   store,
-		RootDir: stateRootDir,
-	})
-}
-
-// InitJobStore must be called after Init
-func (fsr *FsRepo) InitJobStore(prefix string) (jobstore.Store, error) {
-	if exists, err := fsr.Exists(); err != nil {
-		return nil, fmt.Errorf("failed to check if repo exists: %w", err)
-	} else if !exists {
-		return nil, fmt.Errorf("repo is uninitialized, cannot create JobStore")
-	}
-	// load the compute nodes execution store config
-	var storeCfg config_v2.StorageConfig
-	if err := config_v2.GetConfigForKey(config_v2.NodeRequesterJobStore, &storeCfg); err != nil {
-		return nil, err
-	}
-	switch storeCfg.Type {
-	case config_v2.BoltDB:
-		path := storeCfg.Path
-		if path == "" {
-			path = filepath.Join(fsr.path, fmt.Sprintf("%s-requester.db", prefix))
-		}
-		return boltjobstore.NewBoltJobStore(path)
-	case config_v2.InMemory:
-		return memjobstore.NewInMemoryJobStore(), nil
-	default:
-		return nil, fmt.Errorf("unknown JobStore type: %s", storeCfg.Type)
-	}
 }
 
 func (fsr *FsRepo) Init() error {
